Add files via upload
In this web development project I made a music player which is a fun and gratifying project. Features such as volume control and offline listening, as well as the use of different keys such as play/pause and volume, are all taken into account.

this  music player  helped me to  bop on my next web development projects using HTML for functionality, Javascript for audio, keys, and other information, and CSS for styling.